+++ tests/vmss/test_vmss_fetcher.py
@@ -18,7 +18,7 @@ def test_succesful_fetch_vmss(mocked_fetch_vmss):
     result = fetch_vmss(None, None, None)
 
     assert len(result) == 1
-    assert result[0].get('name') == 'chaos-machine'
+    assert result[0].get('name') == 'chaos-pool'
 
 
 @patch('chaosazure.vmss.fetcher.fetch_resources', autospec=True)
@@ -41,8 +41,8 @@ def test_succesful_fetch_instances_without_instance_criteria(mocked_fetch_instan
     result = fetch_instances(scale_set, None, None, None)
 
     assert len(result) == 1
-    assert result[0].get('name') == 'chaos-machine'
-    assert result[0].get('instanceId') == '0'
+    assert result[0].get('name') == 'chaos-pool_0'
+    assert result[0].get('instance_id') == '0'
 
 
 @patch.object(chaosazure.vmss.fetcher, '__fetch_vmss_instances', autospec=True)
@@ -60,84 +60,90 @@ def test_empty_fetch_instances_without_instance_criteria(mocked_fetch_instances)
 def test_succesful_fetch_instances_with_instance_criteria_for_instance0(mocked_fetch_instances):
     # arrange
     instance_0 = vmss_provider.provide_instance()
-    instance_0['instanceId'] = '0'
+    instance_0['instance_id'] = '0'
     instance_1 = vmss_provider.provide_instance()
-    instance_1['instanceId'] = '1'
+    instance_1['instance_id'] = '1'
     instance_2 = vmss_provider.provide_instance()
-    instance_2['instanceId'] = '2'
+    instance_2['instance_id'] = '2'
     instances = [instance_0, instance_1, instance_2]
     mocked_fetch_instances.return_value = instances
     scale_set = vmss_provider.provide_scale_set()
 
     # fire
-    result = fetch_instances(scale_set, [{'instanceId': '0'}], None, None)
+    result = fetch_instances(scale_set, [{'instance_id': '0'}], None, None)
 
     # assert
     assert len(result) == 1
-    assert result[0].get('name') == 'chaos-machine'
-    assert result[0].get('instanceId') == '0'
+    assert result[0].get('name') == 'chaos-pool_0'
+    assert result[0].get('instance_id') == '0'
 
 
 @patch.object(chaosazure.vmss.fetcher, '__fetch_vmss_instances', autospec=True)
 def test_succesful_fetch_instances_with_instance_criteria_for_instance0_instance_2(mocked_fetch_instances):
     # arrange
     instance_0 = vmss_provider.provide_instance()
-    instance_0['instanceId'] = '0'
+    instance_0['instance_id'] = '0'
+    instance_0['name'] = 'chaos-pool_0'
     instance_1 = vmss_provider.provide_instance()
-    instance_1['instanceId'] = '1'
+    instance_1['instance_id'] = '1'
+    instance_1['name'] = 'chaos-pool_1'
     instance_2 = vmss_provider.provide_instance()
-    instance_2['instanceId'] = '2'
+    instance_2['instance_id'] = '2'
+    instance_2['name'] = 'chaos-pool_2'
     instances = [instance_0, instance_1, instance_2]
     mocked_fetch_instances.return_value = instances
     scale_set = vmss_provider.provide_scale_set()
 
     # fire
-    result = fetch_instances(scale_set, [{'instanceId': '0'}, {'instanceId': '2'}], None, None)
+    result = fetch_instances(scale_set, [{'instance_id': '0'}, {'instance_id': '2'}], None, None)
 
     # assert
     assert len(result) == 2
-    assert result[0].get('name') == 'chaos-machine'
-    assert result[0].get('instanceId') == '0'
-    assert result[1].get('name') == 'chaos-machine'
-    assert result[1].get('instanceId') == '2'
+    assert result[0].get('name') == 'chaos-pool_0'
+    assert result[0].get('instance_id') == '0'
+    assert result[1].get('name') == 'chaos-pool_2'
+    assert result[1].get('instance_id') == '2'
 
 
 @patch.object(chaosazure.vmss.fetcher, '__fetch_vmss_instances', autospec=True)
 def test_succesful_fetch_instances_with_instance_criteria_for_all_instances(mocked_fetch_instances):
     # arrange
     instance_0 = vmss_provider.provide_instance()
-    instance_0['instanceId'] = '0'
+    instance_0['instance_id'] = '0'
+    instance_0['name'] = 'chaos-pool_0'
     instance_1 = vmss_provider.provide_instance()
-    instance_1['instanceId'] = '1'
+    instance_1['instance_id'] = '1'
+    instance_1['name'] = 'chaos-pool_1'
     instance_2 = vmss_provider.provide_instance()
-    instance_2['instanceId'] = '2'
+    instance_2['instance_id'] = '2'
+    instance_2['name'] = 'chaos-pool_2'
     instances = [instance_0, instance_1, instance_2]
     mocked_fetch_instances.return_value = instances
     scale_set = vmss_provider.provide_scale_set()
 
     # fire
     result = fetch_instances(
-        scale_set, [{'instanceId': '0'}, {'instanceId': '1'}, {'instanceId': '2'}], None, None)
+        scale_set, [{'instance_id': '0'}, {'instance_id': '1'}, {'instance_id': '2'}], None, None)
 
     # assert
     assert len(result) == 3
-    assert result[0].get('name') == 'chaos-machine'
-    assert result[0].get('instanceId') == '0'
-    assert result[1].get('name') == 'chaos-machine'
-    assert result[1].get('instanceId') == '1'
-    assert result[2].get('name') == 'chaos-machine'
-    assert result[2].get('instanceId') == '2'
+    assert result[0].get('name') == 'chaos-pool_0'
+    assert result[0].get('instance_id') == '0'
+    assert result[1].get('name') == 'chaos-pool_1'
+    assert result[1].get('instance_id') == '1'
+    assert result[2].get('name') == 'chaos-pool_2'
+    assert result[2].get('instance_id') == '2'
 
 
 @patch.object(chaosazure.vmss.fetcher, '__fetch_vmss_instances', autospec=True)
 def test_empty_fetch_instances_with_instance_criteria(mocked_fetch_instances):
     # arrange
     instance_0 = vmss_provider.provide_instance()
-    instance_0['instanceId'] = '0'
+    instance_0['instance_id'] = '0'
     instance_1 = vmss_provider.provide_instance()
-    instance_1['instanceId'] = '1'
+    instance_1['instance_id'] = '1'
     instance_2 = vmss_provider.provide_instance()
-    instance_2['instanceId'] = '2'
+    instance_2['instance_id'] = '2'
     instances = [instance_0, instance_1, instance_2]
     mocked_fetch_instances.return_value = instances
     scale_set = vmss_provider.provide_scale_set()
@@ -145,6 +151,6 @@ def test_empty_fetch_instances_with_instance_criteria(mocked_fetch_instances):
     # fire
     with pytest.raises(FailedActivity) as x:
         fetch_instances(
-            scale_set, [{'instanceId': '99'}, {'instanceId': '100'}, {'instanceId': '101'}], None, None)
+            scale_set, [{'instance_id': '99'}, {'instance_id': '100'}, {'instance_id': '101'}], None, None)
 
         assert "No VMSS instance" in x.value